% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropLazyColumns.R
\name{dropLazyColumns}
\alias{dropLazyColumns}
\title{Drop lazy columns}
\usage{
dropLazyColumns(x)
}
\arguments{
\item{x}{A data.frame containing any number of lazy vectors.}
}
\value{
A data.frame containing only the materialized columns of \code{x}.
}
\description{
Drop columns of a data.frame to only retain non-lazy vectors or lazy vectors that have already been mterialized.
}
\details{
\code{dropLazyColumns} slims down a data.frame of lazy vectors to retain only the columns that were actually used.
The idea is to run this function at the end of a sequence of operations on the data.frames 
produced by \code{\link{createLazyRows}} or \code{\link{createLazyColumns}},
thereby avoiding inadvertent materialization of all columns when saving the workspace.
}
\examples{
Y <- matrix(rnorm(1000), ncol=10)
colnames(Y) <- LETTERS[1:10]

df <- createLazyColumns(Y)
colnames(df)
head(df$A) # used 'A'
tail(df$E) # used 'E'

dropped.df <- dropLazyColumns(df)
colnames(dropped.df)

}
\author{
Aaron Lun
}
