% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropLazyColumns.R
\name{dropLazyColumns}
\alias{dropLazyColumns}
\title{Drop lazy columns}
\usage{
dropLazyColumns(x)
}
\arguments{
\item{x}{A data.frame containing any number of lazy vectors.}
}
\value{
A data.frame containing only the materialized columns of \code{x}.
}
\description{
Drop columns of a data.frame that correspond to yet-to-be-materialized ALTREP vectors.
}
\details{
This function aims to slim down a data.frame of
}
\examples{
Y <- matrix(rnorm(1000), ncol=10)
colnames(Y) <- LETTERS[1:10]

df <- createLazyColumns(Y)
colnames(df)
df$A + 1 # used 'A'
df$E + 2 # used 'E'

dropped.df <- dropLazyColumns(df)
colnames(dropped.df)

}
\author{
Aaron Lun
}
